requirements.txt
opencv-python>=4.10.0
mediapipe>=0.10.14
numpy>=1.26.0
pillow>=10.4.0
matplotlib>=3.9.0
